+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "CGX formatter".
 - Tests for "Genome sequencing".
 - Tests for "Roller coaster".
+- Tests for "Super computer".
 
 ## [1.3.0] - 2022-03-02
 ### Added