+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.4.0] - 2022-03-03
 ### Added
 - Tests for "Blunder - episode 2".
 - Tests for "Blunder - episode 3".
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "TAN network".
 - Tests for "Winamax".
 - Tests for "Music scores".
+- Tests for "The resistance".
 
 ## [1.3.0] - 2022-03-02
 ### Added